# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description: string | null;
  toolCount: number;
}

export interface Tool {
  id: number;
  name: string;
  slug: string;
  category: string;
  description: string;
  /** @nullable */
  pricingStartsAt: string | null;
  pros: string[];
  cons: string[];
  /** @nullable */
  affiliateUrl: string | null;
  /** @nullable */
  logoUrl: string | null;
  featured: boolean;
  isNew: boolean;
  freeTrial: boolean;
  mobileApp: boolean;
  /** @nullable */
  bestFor: string | null;
  integrations: string[];
  /** @nullable */
  rating: string | null;
  createdAt: string;
}

export interface ToolComparison {
  tool1: Tool;
  tool2: Tool;
}

export type DirectoryStatsCategoryBreakdownItem = {
  category: string;
  count: number;
};

export interface DirectoryStats {
  totalTools: number;
  totalCategories: number;
  featuredTools: number;
  categoryBreakdown: DirectoryStatsCategoryBreakdownItem[];
}

export interface SubmitToolBody {
  companyName: string;
  toolName: string;
  category: string;
  website: string;
  contactEmail: string;
  whyFencing: string;
}

export interface ToolSubmission {
  id: number;
  companyName: string;
  toolName: string;
  category: string;
  website: string;
  contactEmail: string;
  whyFencing: string;
  createdAt: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl: string | null;
  publishedAt: string;
}

export interface SubscribeBody {
  email: string;
}

export interface Subscriber {
  id: number;
  email: string;
  createdAt: string;
}

export type ListToolsParams = {
  category?: string;
  search?: string;
  featured?: boolean;
  isNew?: boolean;
  bestFor?: string;
  freeTrial?: boolean;
  mobileApp?: boolean;
};