# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FenceProTools directory API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Tool categories
  - name: tools
    description: Directory tools
  - name: submissions
    description: Tool submission requests
  - name: blog
    description: Blog and resources
  - name: subscribers
    description: Email subscribers
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /categories/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      summary: Get category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found

  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List tools with optional filters
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: isNew
          in: query
          required: false
          schema:
            type: boolean
        - name: bestFor
          in: query
          required: false
          schema:
            type: string
        - name: freeTrial
          in: query
          required: false
          schema:
            type: boolean
        - name: mobileApp
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"

  /tools/featured:
    get:
      operationId: getFeaturedTools
      tags: [tools]
      summary: Get top featured tools (max 6)
      responses:
        "200":
          description: Featured tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"

  /tools/new:
    get:
      operationId: getNewTools
      tags: [tools]
      summary: Get recently added tools
      responses:
        "200":
          description: New tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"

  /tools/{slug}:
    get:
      operationId: getToolBySlug
      tags: [tools]
      summary: Get tool detail by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Not found

  /compare/{slug1}/{slug2}:
    get:
      operationId: compareTools
      tags: [tools]
      summary: Get two tools for side-by-side comparison
      parameters:
        - name: slug1
          in: path
          required: true
          schema:
            type: string
        - name: slug2
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comparison data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolComparison"
        "404":
          description: Not found

  /stats:
    get:
      operationId: getStats
      tags: [tools]
      summary: Directory summary stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryStats"

  /submissions:
    post:
      operationId: submitTool
      tags: [submissions]
      summary: Submit a tool for listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitToolBody"
      responses:
        "201":
          description: Submission received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolSubmission"
        "400":
          description: Validation error

  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog/resource posts
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"

  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

  /subscribe:
    post:
      operationId: subscribe
      tags: [subscribers]
      summary: Subscribe to email newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeBody"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
        "400":
          description: Validation error

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        toolCount:
          type: integer
      required:
        - id
        - name
        - slug
        - description
        - toolCount

    Tool:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        pricingStartsAt:
          type: ["string", "null"]
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        affiliateUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        isNew:
          type: boolean
        freeTrial:
          type: boolean
        mobileApp:
          type: boolean
        bestFor:
          type: ["string", "null"]
        integrations:
          type: array
          items:
            type: string
        rating:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - category
        - description
        - pricingStartsAt
        - pros
        - cons
        - affiliateUrl
        - logoUrl
        - featured
        - isNew
        - freeTrial
        - mobileApp
        - bestFor
        - integrations
        - rating
        - createdAt

    ToolComparison:
      type: object
      properties:
        tool1:
          $ref: "#/components/schemas/Tool"
        tool2:
          $ref: "#/components/schemas/Tool"
      required:
        - tool1
        - tool2

    DirectoryStats:
      type: object
      properties:
        totalTools:
          type: integer
        totalCategories:
          type: integer
        featuredTools:
          type: integer
        categoryBreakdown:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required:
              - category
              - count
      required:
        - totalTools
        - totalCategories
        - featuredTools
        - categoryBreakdown

    SubmitToolBody:
      type: object
      properties:
        companyName:
          type: string
        toolName:
          type: string
        category:
          type: string
        website:
          type: string
        contactEmail:
          type: string
        whyFencing:
          type: string
      required:
        - companyName
        - toolName
        - category
        - website
        - contactEmail
        - whyFencing

    ToolSubmission:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        toolName:
          type: string
        category:
          type: string
        website:
          type: string
        contactEmail:
          type: string
        whyFencing:
          type: string
        createdAt:
          type: string
      required:
        - id
        - companyName
        - toolName
        - category
        - website
        - contactEmail
        - whyFencing
        - createdAt

    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        publishedAt:
          type: string
      required:
        - id
        - title
        - slug
        - excerpt
        - content
        - imageUrl
        - publishedAt

    SubscribeBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email

    Subscriber:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - createdAt