# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  description: zod.string().nullable(),
  toolCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get category by slug
 */
export const GetCategoryBySlugParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetCategoryBySlugResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  description: zod.string().nullable(),
  toolCount: zod.number(),
});

/**
 * @summary List tools with optional filters
 */
export const ListToolsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  featured: zod.coerce.boolean().optional(),
  isNew: zod.coerce.boolean().optional(),
  bestFor: zod.coerce.string().optional(),
  freeTrial: zod.coerce.boolean().optional(),
  mobileApp: zod.coerce.boolean().optional(),
});

export const ListToolsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  category: zod.string(),
  description: zod.string(),
  pricingStartsAt: zod.string().nullable(),
  pros: zod.array(zod.string()),
  cons: zod.array(zod.string()),
  affiliateUrl: zod.string().nullable(),
  logoUrl: zod.string().nullable(),
  featured: zod.boolean(),
  isNew: zod.boolean(),
  freeTrial: zod.boolean(),
  mobileApp: zod.boolean(),
  bestFor: zod.string().nullable(),
  integrations: zod.array(zod.string()),
  rating: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListToolsResponse = zod.array(ListToolsResponseItem);

/**
 * @summary Get top featured tools (max 6)
 */
export const GetFeaturedToolsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  category: zod.string(),
  description: zod.string(),
  pricingStartsAt: zod.string().nullable(),
  pros: zod.array(zod.string()),
  cons: zod.array(zod.string()),
  affiliateUrl: zod.string().nullable(),
  logoUrl: zod.string().nullable(),
  featured: zod.boolean(),
  isNew: zod.boolean(),
  freeTrial: zod.boolean(),
  mobileApp: zod.boolean(),
  bestFor: zod.string().nullable(),
  integrations: zod.array(zod.string()),
  rating: zod.string().nullable(),
  createdAt: zod.string(),
});
export const GetFeaturedToolsResponse = zod.array(GetFeaturedToolsResponseItem);

/**
 * @summary Get recently added tools
 */
export const GetNewToolsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  category: zod.string(),
  description: zod.string(),
  pricingStartsAt: zod.string().nullable(),
  pros: zod.array(zod.string()),
  cons: zod.array(zod.string()),
  affiliateUrl: zod.string().nullable(),
  logoUrl: zod.string().nullable(),
  featured: zod.boolean(),
  isNew: zod.boolean(),
  freeTrial: zod.boolean(),
  mobileApp: zod.boolean(),
  bestFor: zod.string().nullable(),
  integrations: zod.array(zod.string()),
  rating: zod.string().nullable(),
  createdAt: zod.string(),
});
export const GetNewToolsResponse = zod.array(GetNewToolsResponseItem);

/**
 * @summary Get tool detail by slug
 */
export const GetToolBySlugParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetToolBySlugResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  category: zod.string(),
  description: zod.string(),
  pricingStartsAt: zod.string().nullable(),
  pros: zod.array(zod.string()),
  cons: zod.array(zod.string()),
  affiliateUrl: zod.string().nullable(),
  logoUrl: zod.string().nullable(),
  featured: zod.boolean(),
  isNew: zod.boolean(),
  freeTrial: zod.boolean(),
  mobileApp: zod.boolean(),
  bestFor: zod.string().nullable(),
  integrations: zod.array(zod.string()),
  rating: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Get two tools for side-by-side comparison
 */
export const CompareToolsParams = zod.object({
  slug1: zod.coerce.string(),
  slug2: zod.coerce.string(),
});

export const CompareToolsResponse = zod.object({
  tool1: zod.object({
    id: zod.number(),
    name: zod.string(),
    slug: zod.string(),
    category: zod.string(),
    description: zod.string(),
    pricingStartsAt: zod.string().nullable(),
    pros: zod.array(zod.string()),
    cons: zod.array(zod.string()),
    affiliateUrl: zod.string().nullable(),
    logoUrl: zod.string().nullable(),
    featured: zod.boolean(),
    isNew: zod.boolean(),
    freeTrial: zod.boolean(),
    mobileApp: zod.boolean(),
    bestFor: zod.string().nullable(),
    integrations: zod.array(zod.string()),
    rating: zod.string().nullable(),
    createdAt: zod.string(),
  }),
  tool2: zod.object({
    id: zod.number(),
    name: zod.string(),
    slug: zod.string(),
    category: zod.string(),
    description: zod.string(),
    pricingStartsAt: zod.string().nullable(),
    pros: zod.array(zod.string()),
    cons: zod.array(zod.string()),
    affiliateUrl: zod.string().nullable(),
    logoUrl: zod.string().nullable(),
    featured: zod.boolean(),
    isNew: zod.boolean(),
    freeTrial: zod.boolean(),
    mobileApp: zod.boolean(),
    bestFor: zod.string().nullable(),
    integrations: zod.array(zod.string()),
    rating: zod.string().nullable(),
    createdAt: zod.string(),
  }),
});

/**
 * @summary Directory summary stats
 */
export const GetStatsResponse = zod.object({
  totalTools: zod.number(),
  totalCategories: zod.number(),
  featuredTools: zod.number(),
  categoryBreakdown: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Submit a tool for listing
 */
export const SubmitToolBody = zod.object({
  companyName: zod.string(),
  toolName: zod.string(),
  category: zod.string(),
  website: zod.string(),
  contactEmail: zod.string(),
  whyFencing: zod.string(),
});

/**
 * @summary List blog/resource posts
 */
export const ListBlogPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  imageUrl: zod.string().nullable(),
  publishedAt: zod.string(),
});
export const ListBlogPostsResponse = zod.array(ListBlogPostsResponseItem);

/**
 * @summary Get a blog post by slug
 */
export const GetBlogPostParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetBlogPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  excerpt: zod.string(),
  content: zod.string(),
  imageUrl: zod.string().nullable(),
  publishedAt: zod.string(),
});

/**
 * @summary Subscribe to email newsletter
 */
export const SubscribeBody = zod.object({
  email: zod.string(),
});