# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl: string | null;
  publishedAt: string;
}