# lib/api-zod/src/generated/types/directoryStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */
import type { DirectoryStatsCategoryBreakdownItem } from "./directoryStatsCategoryBreakdownItem";

export interface DirectoryStats {
  totalTools: number;
  totalCategories: number;
  featuredTools: number;
  categoryBreakdown: DirectoryStatsCategoryBreakdownItem[];
}