# lib/api-zod/src/generated/types/subscriber.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */

export interface Subscriber {
  id: number;
  email: string;
  createdAt: string;
}