# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */

export interface Tool {
  id: number;
  name: string;
  slug: string;
  category: string;
  description: string;
  /** @nullable */
  pricingStartsAt: string | null;
  pros: string[];
  cons: string[];
  /** @nullable */
  affiliateUrl: string | null;
  /** @nullable */
  logoUrl: string | null;
  featured: boolean;
  isNew: boolean;
  freeTrial: boolean;
  mobileApp: boolean;
  /** @nullable */
  bestFor: string | null;
  integrations: string[];
  /** @nullable */
  rating: string | null;
  createdAt: string;
}