# lib/api-zod/src/generated/types/toolComparison.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */
import type { Tool } from "./tool";

export interface ToolComparison {
  tool1: Tool;
  tool2: Tool;
}