# lib/api-zod/src/generated/types/toolSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FenceProTools directory API
 * OpenAPI spec version: 0.1.0
 */

export interface ToolSubmission {
  id: number;
  companyName: string;
  toolName: string;
  category: string;
  website: string;
  contactEmail: string;
  whyFencing: string;
  createdAt: string;
}